From: : " Michael"
packet-tcp marks a row as unwritable if there is more
than 1 PDU in a packet to prevent overwriting of the
protocol column.

However this breaks the proto,colinfo tap which checks
rows are writable or dies EVEN if it doesn't alter anything.

Suggested patch attached to restore the initial value
after all PDUs have been processed.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3675

svn path=/trunk/; revision=29535