From Mithun Roy: The primary header in bundle protocol contains some offset values(destination scheme offset, destination ssp offset, source scheme offset, source ssp offset, etc). These are the offsets within the dictionary if the length of the dictionary length is greater than 0. But if the dictionary length is 0, then these offset refer to node number and service number respectively(according to compressed bundle header encoding). For example if destination scheme offset is 2 and the destination ssp offset is 1, then the destination EID(<node_number>.<service_number>) is 2.1.
Currently the dtn dissector will consider these offsets to be actual offsets in
the dictionary even if the dictionary length is 0. So the values for the
EID's(destination, source, report, custodian) and their schemes are junk
values. For example if the destination scheme offset is 2 and the destination
ssp offset is 1 and the dictionary length is 0(which means the dictionary is
empty), then the destination scheme is 2 bytes after the beginning of the
metadata block(field after dictionary) and destination is 1 byte after the
beginning of the metadata block.

svn path=/trunk/; revision=30682