Copy over with manual intervention: ------------------------------------------------------------------------ r32396 | guy | 2010-04-05 19:49:18 -0700 (Mon, 05 Apr 2010) | 4 lines Changed paths:    M /trunk/plugins/docsis/packet-bpkmrsp.c
The third argument to val_to_str() is a format string that gets passed
the value that couldn't be mapped to a string - it's an integer and
needs to be printed with an integer format.  Fixes bug 4644.
------------------------------------------------------------------------
r32398 | morriss | 2010-04-06 06:40:06 -0700 (Tue, 06 Apr 2010) | 1 line
Changed paths:
   M /trunk/plugins/docsis/packet-regrsp.c

Don't use %s as the format string in val_to_str().  This fixes the fuzz failure in bug 4646.
------------------------------------------------------------------------

Update the release notes.

svn path=/trunk-1.0/; revision=32662