From Chris Maynard via bug 4722:
When I submitted the USB and PPP changes, I had not previously tested with a
full insertion of a particular USB device.  Now that I have, I've found that
what used to be reported as an unknown interface class is now reported as a
vendor specific interface class.  Thus, in order for PPP/USB dissection to
work, the ppp dissector must register for that class as well.  The attached
trivial patch does just that.

svn path=/trunk/; revision=32768