From Chris Maynard: Fix for bug 4807 to handle compressed (NULL) address and control bytes.
The problem seems to be that the address and control bytes are missing,
presumably because they are compressed per section 3.2 of
http://tools.ietf.org/html/rfc1662.  The attached patch tries to account for
that.

The patch also fixes another bug.  If the hdlc framing character, 0x7e, is not
present, but the address and control fields are, then the offset might need to
be 3 bytes instead of only 2 bytes if the control field is escaped.  Sometimes
it is escaped and sometimes it isn't, so we need to handle both cases.

svn path=/trunk/; revision=33021