From Jason Burns:  DHCP Dissector - Support for DHCP Option 150 TFTP Server Address
"According to this document DHCP Option 150 is an array of IPv4 TFTP Server
Addresses

http://cisco.com/en/US/products/sw/voicesw/ps556/products_tech_note09186a00800942f4.shtml#tftp_option

I fixed the capitalization of this option in the bootp dissector and changed
the type to match the actual use of the option. (ipv4_list)"

From me: Add a pointer to a draft RFC about the DHCP "TFTP Server Address" option.



svn path=/trunk/; revision=33203