Pick up fix from the main branch:
r33572 | guy | 2010-07-18 12:41:11 -0700 (Sun, 18 Jul 2010) | 8 lines

Initial support for the trailer information in versions 2.1 and later of
the NetMon file format.  Currently, we just use the network type field,
and we ignore all the special record types and don't try to handle any
of the other special network types.

We also catch bogus frame tables where the record is bigger than the
frame table says it is.

svn path=/trunk-1.4/; revision=33573