Pick up from the trunk:
r33613 | guy | 2010-07-22 02:11:16 -0700 (Thu, 22 Jul 2010) | 8 lines

In the frame_data structure, expand the per-packet
encapsulation/data-link type to 16 bits, and shuffle some fields to
eliminate some unnecessary padding - the net result should be no change
in the structure size for 32 bits and a few bytes removed for 64 bits.

This allows more encapsulation types - we've just about run out of the
ones that fit in a signed 8-bit integer - and thus should fix bug 5025.

svn path=/trunk-1.4/; revision=33614