Copy over r33531.
------------------------------------------------------------------------
r33531 | guy | 2010-07-14 19:07:16 -0700 (Wed, 14 Jul 2010) | 9 lines
Changed paths:
   M /trunk/Makefile.am
   M /trunk/epan/epan.c
   M /trunk/epan/epan.h
   M /trunk/epan/libwireshark.def
   M /trunk/gtk/main.c
   M /trunk/rawshark.c
   M /trunk/tshark.c
   M /trunk/version_info.c
   M /trunk/version_info.h

Move the code to get version information for libraries used by
libwireshark into libwireshark, and call it only in programs linked with
libwireshark.  That way, programs that don't link with libwireshark
don't have to link with libgcrypt or libgnutls solely so that they can
say that they're linked with a particular version of libgcrypt or
libgnutls.

Don't link dumpcap with libgcrypt or libgnutls any more.
------------------------------------------------------------------------

svn path=/trunk-1.4/; revision=33999