From Jakub Zawadzki via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3336 :
We parse host & request page from headers, so we easily can construct full http
uri.

I was thinking about making it as field, so we could filter, print in column
info, or do other fancy stuff, but for now this is imho enough.


From me: add it as a (filterable) item.  Clean up spacing and indentation in a
few places.

svn path=/trunk/; revision=34162