Generate the value_string array sminmpec_values[] with "gaps" filled in (using name "(Unknown)" so that direct (indexed) access to the value_string is possible.
Note that this represents a slight API change in that
match_strval() on sminmpec_values  will no longer return NULL for
missing enterprise numbers (less than the maximum in the array)
but will return the string "(Unknown)".

AFAIKT this change does not materially affect existing Wireshark
use of the sminmpec value_string array.

svn path=/trunk/; revision=34218