From Colin O'Flynn and Owen Kirby via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5266 :
The attached patch begins to add support for RPL to the ICMPv6 file. All
locations that RPL code have been added are marked with a comment allowing this
patch to be reverted at a future time if it is decided to e.g. move all the RPL
code to it's own dissector.

A few values await IANA assignment and are also clearly marked (in
packet-ipv6.h). 

Only the 'metric' option is left unsupported, as it is primarily defined in
another I-D.

svn path=/trunk/; revision=34579