Move main_statusbar.h to the top-level directory; none of its routines use GTK+ data types, so, at least in theory, it could be implemented atop another toolkit.
Make statusbar_push_temporary_msg() take a format string and format
arguments.  Use it instead of simple_status(), and change one call to
just take a format string and arguments rather than to take the result
of using that format string and arguments with g_strdup_printf() and
passing the result to statusbar_push_temporary_msg().

svn path=/trunk/; revision=35041