From Neil Piercy via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=3301 :
The packet-sccp.c has a bug in the declared valid ranges of the SSN and DPC
values in the user table used to match to a subdissector. The SSN range is 16
bits rather than 8 (not really an issue) but the DPC range is 16 bits rather
than 24 - so many traces cannot be matched by this table.

svn path=/trunk/; revision=35097