From Zachary Mark via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5457 :
The scsi_persresv_type_val field in packet-scsi.c contains a mapping of
persistent reservation opcodes to their descriptive types.  The opcode for the
Exclusive Access - Registrants Only field is incorrectly set to 7, when the
correct opcode is 6 (as per SPC-2 onward).  The attached patch corrects this
discrepancy.

The attached patch also adds support for dissecting opcodes 7 and 8, the two
all registrants reservation types present in SPC-3 onward.

svn path=/trunk/; revision=35099