From daulis0@gmail.com: There are 2 parts to the CIGI protocol. 1) Host to IG messages and 2) IG to Host messages. Currently, Host to IG messages are parsed correctly, but IG to Host messages show Malformed Packet (or may not even detect as CIGI at all)
Some of the protocol format is different between versions. The "Minor Version"
is used by the dissector to separate the differences, but this field is in a
different location in the IG Control Packet (Host to IG message) vs the Start
of Frame Packet (IG to Host message).

Attached patch to correct this.

svn path=/trunk/; revision=35241