When generating a tvb subset for chunks, parameters, and error causes, set the:  - "reported length" to be minimum of the entity's length (taken from the    packet) or the reported length of the TVB.  - "backing length" to be the minimum of the reported length (above) or the    (captured) length of the TVB.
This prevents tvb_new_subset() from generating an exception if the entity's
length (in the packet) is bogus (bigger than what's in the TVB) which allows
dissection to continue to a point where we can show the user what the problem
is.

When dissecting chunks, add an expert info if the item's length is bigger than
what was on the wire (the reported length).  (The same could also be done for
parameters and error causes.)

This makes captures like that in https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5546
easier to understand.

It also starts getting this dissector using the reported length instead of the
(captured) TVB length (there's more to be done with other tvb subsets).

svn path=/trunk/; revision=35366