Copy over revisions from the trunk:
------------------------------------------------------------------------
r35295 | jake | 2010-12-29 10:16:50 -0800 (Wed, 29 Dec 2010) | 2 lines
Changed paths:
   M /trunk/docbook/wsug_src/WSUG_chapter_capture.xml

Update capture options dialog description.
------------------------------------------------------------------------
r35304 | guy | 2010-12-29 18:56:51 -0800 (Wed, 29 Dec 2010) | 9 lines
Changed paths:
   M /trunk/epan/dissectors/packet-dcerpc-samr.c
   M /trunk/epan/dissectors/pidl/samr.cnf
   M /trunk/epan/dissectors/pidl/samr.idl

The rids argument to samr_GetGroupsForUser() is *not* a single RID, it's
an array of samr_RidWithAttributeArray structures.  Don't equate it to
hf_samr_rid; that causes hf_samr_rid to be added as an item with a
length of -1, but, as hf_samr_rid is an FT_UINT32 field, that causes a
dissector bug error.

The version of Yapp I was using, at least, didn't like C++-style
comments in the IDL file; replace them with C-style comments.
------------------------------------------------------------------------
r35313 | guy | 2010-12-30 16:02:54 -0800 (Thu, 30 Dec 2010) | 4 lines
Changed paths:
   M /trunk/epan/dissectors/packet-tcp.c

It's not a good idea to use tcph->th_seglen if you haven't set it.

Fixes bug 4211.
------------------------------------------------------------------------
r35322 | stig | 2011-01-01 04:23:21 -0800 (Sat, 01 Jan 2011) | 2 lines
Changed paths:
   M /trunk/packaging/macosx/Info.plist.in
   M /trunk/version_info.c

New Year.
------------------------------------------------------------------------
r35328 | guy | 2011-01-02 11:39:24 -0800 (Sun, 02 Jan 2011) | 2 lines
Changed paths:
   M /trunk/plugins/unistim/packet-unistim.c

Fix some cut-and-pasteos and other typoes.
------------------------------------------------------------------------
r35332 | guy | 2011-01-02 19:30:20 -0800 (Sun, 02 Jan 2011) | 3 lines
Changed paths:
   M /trunk/epan/dissectors/packet-diameter.c

Do length checking for Integer32, Integer64, Unsigned64, Float32, and
Float64, as well as for Unsigned32.
------------------------------------------------------------------------
r35333 | jake | 2011-01-02 23:40:21 -0800 (Sun, 02 Jan 2011) | 2 lines
Changed paths:
   M /trunk/docbook/wsdg_graphics/ws-function-blocks.dia
   M /trunk/docbook/wsdg_graphics/ws-function-blocks.png
   M /trunk/docbook/wsdg_src/WSDG_chapter_works.xml

Include dumpcap in the description.
------------------------------------------------------------------------
r35335 | guy | 2011-01-03 00:21:35 -0800 (Mon, 03 Jan 2011) | 4 lines
Changed paths:
   M /trunk/epan/dissectors/packet-diameter.c

If we have no value to dissect, because the AVP has the wrong length for
the value's type, return NULL as the label, rather than returning an
uninitialized value as the label.
------------------------------------------------------------------------
r35336 | guy | 2011-01-03 00:31:45 -0800 (Mon, 03 Jan 2011) | 6 lines
Changed paths:
   M /trunk/epan/dissectors/packet-tn3270.c

Put the query list items under the top-level query list item (which
should've been done in the previous checkin).

Display the "request type" field in a Read Partition request correctly
(only the upper 2 bits matter, and they're an enumerated value).
------------------------------------------------------------------------
r35337 | guy | 2011-01-03 01:17:20 -0800 (Mon, 03 Jan 2011) | 9 lines
Changed paths:
   M /trunk/wiretap/vms.c

We must always return an error code on an error; otherwise, our caller
will see random crap as the error code.

However, if we're skipping a "TCPIPTRACE-W-BUFFERSFUL" error, if the
"error" we get is an end-of-file indication, that's *not* an error.

It is, however, ultimately a "we dropped some packets" indication; add a
comment noting that we should eventually treat it as such.
------------------------------------------------------------------------
r35356 | guy | 2011-01-03 23:48:28 -0800 (Mon, 03 Jan 2011) | 9 lines
Changed paths:
   M /trunk/epan/dissectors/packet-ltp.c

Instead of going over the Report segment twice, once to get the fields
in the segment and once to put them into the protocol tree, just do it
once.  That also means we don't need to allocate an array for all the
reception claims, so we don't have to worry about the claim count (other
than making sure it's non-negative), and that we won't abort the
dissection until we run past the end of the packet.

(The rest of the dissector should be changed to work that way as well.)
------------------------------------------------------------------------
r35357 | guy | 2011-01-04 00:04:27 -0800 (Tue, 04 Jan 2011) | 3 lines
Changed paths:
   M /trunk/epan/dissectors/packet-amqp.c

Don't use DISSECTOR_ASSERT() to catch malformed packets.  Fixes bug
4048.
------------------------------------------------------------------------
r35369 | guy | 2011-01-04 22:11:10 -0800 (Tue, 04 Jan 2011) | 2 lines
Changed paths:
   M /trunk/epan/dissectors/packet-amqp.c

Put end-of-field-table checks back into dissect_amqp_field_table().
------------------------------------------------------------------------


Copy over by hand:
------------------------------------------------------------------------
r35330 | guy | 2011-01-02 13:27:57 -0800 (Sun, 02 Jan 2011) | 12 lines
Changed paths:
   M /trunk/epan/dissectors/packet-ber.c

For those cases where we interpret the contents of an OCTET STRING,
which could be of arbitrary length - even if it's not supposed to be! -
as a value of some other type, by adding them as a registered field,
first check to make sure the length of the field is appropriate for the
type and, if not, show a dissection error, rather than showing a
dissector-bug assertion when we call proto_tree_add_item().

This fixes a bunch of dissector-bug assertions that show up with
malformed BER-encoded packets.

Also, fix a typo, and expand a comment.
------------------------------------------------------------------------

Update the release notes.

svn path=/trunk-1.4/; revision=35462