From Johannes Lange  via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5606 :
In semcheck.c the display filter string of an expression is checked against the
header_field_info.display value BASE_CUSTOM. But the value of BASE_CUSTOM is
applied as bitmask while the actual type is an enum (BASE_CUSTOM = 6).

With this BASE_DEC, BASE_DEC_HEX and BASE_HEX_DEC are also matching and are not
accepted as filter expression.


Actually: BASE_DEC works but not BASE_HEX.  And the problem only shows up when
trying to match a field in one of these bases against a string (from a
value_string).

svn path=/trunk/; revision=35621