Copy over revisions from the trunk:
------------------------------------------------------------------------
r36101 | gerald | 2011-02-28 16:50:17 -0800 (Mon, 28 Feb 2011) | 5 lines
Changed paths:
   M /trunk/asn1/ldap/ldap.cnf
   M /trunk/asn1/ldap/packet-ldap-template.c
   M /trunk/epan/dissectors/packet-ldap.c

If we encounter an excessively long filter string or one with too many
elements, add an expert item and throw a ReportedBoundsError. We may
want to handle this more cleanly in the future but this avoids
allocating a huge amount of memory.
------------------------------------------------------------------------

Update the release notes.

svn path=/trunk-1.4/; revision=36102