Based on https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5821#c2 :
Move the setting of debug environment variables out of the while loop.

Export MALLOC_OPTIONS=AJ to increase memory allocation debug on FreeBSD (and
others).

Export a number of environment variables that MacOS looks at.

(I don't have easy access to either of these OS so this has not been tested.)

Add a comments for each entry explaining which OS uses it and what it does.

svn path=/trunk/; revision=36622