From Roland Knall via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5786 :
This patch registers a dissector table to allow other dissectors to register as
subdissectors for modbus/tcp data.

Additionally it registers the dissector, so it can be invoked by another
dissector by name.

Additionally this version of the patch adds the Data field not as a text field,
but rather as a field which can be selected and tested for by name.

svn path=/trunk/; revision=36925