SVN #37419 fixed an OSX compile issue and reverted SVN #37408.
The actual cause of an OSX compile problem: The text   [^[:blank:]]   in a comment.

SVN #37408 was reverted because a period (or most other characters)
in a RE character class don't need to be escaped.

This patch adds some replacement text in the spot in the comment
originally occupied by the text    [^[:blank:]]


svn path=/trunk/; revision=37421