From Michael Mann: UATification of ESP
UATification of ESP SA keys.  Some notes:
1. Trimmed a lot of fat removing the code that parsed (and reparsed) ESP SA
keys, since UATs can take care of all that. 
2. There was no validation on the src/dest, so that was carried through (ie
there is still no validation of src/dest fields)
3. Formatting of the source was a little off, so I attempted to fix the
functions I needed to deal with.  I know that's frowned upon, but the braces
were such that I couldn't tell if/else blocks to understand the code.  If
somebody wanted to run an indent program on the file, that would be
appreciated.

Added some casts to make it compile on win64.
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=1123

svn path=/trunk/; revision=37516