Don't return an offset that you won't later use.
The length fields in a pcap_pkthdr are unsigned, so presumably the
equivalent fields in the rpcap protocol are also unsigned.

Make sure the captured data length isn't bigger than the remaining data
in the packet before attempting to construct a tvbuff for the packet
data.  If it is, report that as an error, and don't even try to
construct the tvbuff; that'll fail.  This fixes bug 6073.

svn path=/trunk/; revision=37826