From Christopher Maynard: From doc/README.tapping:
IF the tap is going to return private data using the last parameter to
tap_queue_packet() and IF the protocol can appear multiple times inside the
same packet, you will have to make sure that each instance of
tap_queue_packet() is using its own instance of private struct variable
so they don't overwrite each other.

This was the case here, so I changed the static allocation of tap_rec to be
ep_alloc'd.  I also tried to clean up the Info column a bit.  With this patch
applied, the graph now shows the APM and ANM correctly instead of 2 ANM's.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5966

svn path=/trunk/; revision=37893