From Dirk Jagdmann via bug 6194:
The attachted patch fixes and enhances the SCPS TCP option dissection. Changes
are:

- fix order of reserved Bit 1,2,3
- fix minimum TCP option length
- fix proto items
- add proto item for Connection ID
- removed the verify_scps() function. It's logic was broken, because it did
reset the scps_capable flag on both flows if one of them did not have it.
However sometimes that flag is only enabled in one flow direction and that flow
direction could see TCP options later on, which would get dissected as invalid.
See the attachted capture file for an example.

svn path=/trunk/; revision=38326