From Anders on -dev:
Brutal patch to build with GTK 3.0 

With the enclosed patch Wireshark builds on GTK3.0. Colors are messed up etc but
it should be possible to start to fix things and actually see what's happening.
I wouldn't be surprised if many patches leading up to this are faulty. Color
handling should probably be rewritten to use GdkRGBA in many places etc.

  http://www.wireshark.org/lists/wireshark-dev/201108/msg00383.html 


svn path=/trunk/; revision=38612