From Pierre-Marie de Rodat: PostgreSQL Startup message not properly supported by the PostgreSQL dissector.
"The PostgreSQL dissector do not fully support the frontend StartupMessage (see
"StartupMessage" in
http://developer.postgresql.org/pgdocs/postgres/protocol-message-formats.html).
The couples parameter name/parameter value in this kind of message are reported
as a block of text ("name: value") by the dissector whereas reporting them as
parameter name/parameter value would be more appropriate.

I've fixed it, so now the username and the database sent by the frontend can be
handled in, for instance, the CSV output of TShark.

I've also added a "val_count" field to contain the number of values (row
descriptions or row data) included in RowDescription/DataRow messages. This
information is useful when analyzing the CSV of TShark since in a CSV row, many
row descriptions or row data may be packed together."


Patch changes from me:
- No need to fetch ephemeral string anymore so just use tvb_strsize()
  to get string length;
- Change field-filtername from pgsql.val.count to pgsql.field.count

See: https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6343


svn path=/trunk/; revision=39030