Clean up white space.
The appropriate encoding for a string is ENC_BIG_ENDIAN or
ENC_LITTLE_ENDIAN, combined with a character encoding value; it is not
ENC_NA, as the character encoding is definitely applicable, and the byte
order is also applicable if it's a counted string or a UTF-16 or UCS-2
string.  Assume ASCII for now.

svn path=/trunk/; revision=39187