As bug 3954 notes, nowhere does RFC 3954 indicate that a system or interface scope is always 4 bytes.  For that matter, nowhere does it indicate that the scopes have any particular interpretation except as a sequence of octets.
Get rid of the checks for a length of 4, and make ScopeSystem an
FT_BYTES.  If, by *convention*, they're usually IPv4 or IPv6 addresses,
somebody can throw in code to display them as such if they happen to be
4 or 16 bytes, respectively.  Leave ScopeInterface as an integer for
now, in case, by convention, they're interface indices, but still leave
the length check out.

Fixes bug 3954.

svn path=/trunk/; revision=39485