From Lori Tribble:
I found many new properties in BACnet that were not decoded correctly in
Wireshark.  I have attached a patch that adds decoding to the following
properties:

Fixed the following properties to decode:

In the Accumulator Object:
scale (187)
prescale (185)
logging-record (184)

In Access Door Object:
door-members (228)
masked-alarm-values (234)

In the Pulse Converter Object:
input-reference (181)

In the Group Object:
list-of-group-members (53)

In the Event Enrollment Object:
object-property-reference (78)

In the Command Object:
action (2)

In the Trend Log Multiple object:
log-buffer (131)

In the Event Log Object:
log-buffer (131)

In the Structured View object:
subordinate-list (211)

Added decoding for the following enumeration type properties:
  Logging-type (197)
  event-state (36)
  reliability (103)
  notify-type (208)
  door-status (231)
  lock-status (233)
  secured-status (235)
  maintenance-required (158)
  program-state (92)
  program-change (90)
  reason-for-halt (100)
  mode (160)
  silenced (163)
  operation-expected (161)
  tracking-value (164)
  file-access-method (41)

The above fixes also fixed the log-buffer decoding of the old trend log 
recorded as Wireshark bug: #6458

I believe this patch also fixes bug: #6235.  But the original bug did not
contain a capture file so I am not positive my capture represents the
originators issue.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6475

svn path=/trunk/; revision=39676