Clean up indentation.
Do more length checks, so we reject trailers that have nothing to
dissect.

Test for the trailer length being >= 8, rather than for having the 0x08
bit set (they amount to the same thing, as the trailer length is <= 14,
and >= 8 is what we really want if we're checking whether there's a
timestamp).

svn path=/trunk/; revision=40142