In a loop of the form
	for (i = 1; i <= N; i++)
		...

the type of "i" must have, as its maximum value, a value >= the maximum
value of N; otherwise, if N is equal to the maximum value that fits in
"i", the loop willnever terminate.  (If that requires "i" to be larger
than you'd like, do the loop as

	for (i = 0; i < N; i++)
		...

which doesn't have that problem.)

Clean up the "i = 1" clause's white space in those for loops.

svn path=/trunk/; revision=41010