(Re)fix https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6712 :
The change made in r40742 turned out to not be portable (FreeBSD doesn't have
the timezone global variable), so use another method to determine the current
timezone.

Also fix a bug introduced by r40742's change to display this timestamp in UTC:
if the reported (by the message) time zone has a negative offset to UTC, shift
it forward (not backward) to get UTC (and the opposite for positive offsets).

svn path=/trunk/; revision=41044