From Martin Kaiser via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6890 :
The attached patch for the DVB-CI dissector creates a circuit for each DVB-CI
session, using the session number as circuit id.

The DVB-CI session commands are:

- open_session_request(resource_id)
- session_opened(resource_id, newly assigned session number)
- payload transfer(session_number, payload data)
- close session(session_number)

For now, the circuit will store the resource id and make it available (as a
generated item) to subsequent packets that contain only the session number.
Doing this, the resource id (which is like a tcp/udp port) can be used for
filtering.

svn path=/trunk/; revision=41253