From Weston Andros Adamson via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6919 :
Both deviceid4 and sessionid4 are currently decoded as rpc_opaque_data - this
means their values are never displayed in the "Packet Detail" section and to
see them you have to select the name in the "Packet Detail" section and look at
the byte range in the "Packet Bytes" section.

The attached patch (against today's SVN) adds dissectors for both so these
values are displayed in the "Packet Detail" section.

svn path=/trunk/; revision=41417