Have AC_WIRESHARK_GCC_CFLAGS_CHECK take an optional second flag to indicate whether the flag is C-only, C++-only, or for both compilers; pass the appropriate value for C-only flags.  Have the "Checking for..." message indicate whether we're adding to CFLAGS, CXXFLAGS, or both.
(Yes, the macro should probably be renamed.  "GCC" refers to the GNU
Compiler Collection, which includes a C++ compiler, although that's also
used for compilers that are more-or-less compatible with the ones from
the GNU Compiler Collection, such as the clang compilers.)

We set -Wformat-security whether or not --enable-extra-gcc-checks was
specified, so we don't need to do it again if it was specified.

svn path=/trunk/; revision=41586