Get the Lua menus working again: when adding UI menus, it appears that our description of existing menus must be accurate: if a name and action are both specified then the to-be-added-XML must include both the name and the action.
It appears that the formats given in stat_group_name()'s value_string were
designed to take this into account, but there was no code to separate the menu
name from its action.  Adding that separation is complicated by the fact that
menus are separated by /'s and actions may also contain that character.  To
deal with that, "escape" the /'s in actions by replacing them with #'s.
make_menu_xml() then un-escapes them back into /'s.

There has to be an easier way to do all of this...

svn path=/trunk/; revision=41591