From Martin Kaiser via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7023 :
For each message, the DVB-CI dissector checks the message size against the
buffer size negotiated at startup. If the buffer size negotation was not
captured, a warning is flagged up for each packet saying its size is too big. 

Obviously, the check only makes sense if we know the negotiated buffer size...

svn path=/trunk/; revision=41901