From Metatech via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7014
Here is a dissector for ActiveMQ OpenWire protocol.

A few words about the protocol : 
OpenWire has two wire formats :
- "loose" : more verbose, less CPU-intensive, less network-intensive (1-pass)
- "tight" : more compact, more CPU-intensive, more network-intensive (2-pass)
This dissector only supports the "loose" syntax, which is not the default.
This dissector only supports version 6 of the protocol.
It can be changed on the broker in the activemq.xml file by specifying
"tightEncodingEnabled=false" :


svn path=/trunk/; revision=41919