Fix the name of the variable with the handle for the V5.2 dissector, to make it clearer that the payload isn't being dissected as uninterpreted data.
Whether the checksum is checked should not depend n the presence of
per-frame protocol data; make it depend on the availability of the
checksum, instead.  Then #if 0 out all that code, as, at least in the
V5.2 capture attached to bug 7027, there is no checksum in any of the
packets.

Clean up indentation a bit.

svn path=/trunk/; revision=41963