If we don't even *have* a filter combo box, there's nothing we can check or colorize; quit before we even *try* to get the link-layer header type.
If we don't have an active link-layer header type, that probably means
we don't know what link-layer header types the interface provides, which
would be the case if it were a named pipe or an interface we can't open.
Don't crash in that case, just leave the filter combo box uncolored, to
indicate that we have no clue whether the filter is valid or not.

Fixes bug 7049.

svn path=/trunk/; revision=42059