Get the filter text box before we even look at the link-layer header types, and bail if we dn't have one.
Fix the comment for the "selected link-layer header type is unsupported"
case, and colorize the text box as empty in that case.

Use "Unknown" rather than "not supported" for the dummy link-layer
header type entry we add if we don't have a list of link-layer header
types; at least for pipes, we don't, and *can't*, know what link-layer
header type we'll see.

svn path=/trunk-1.6/; revision=42075