From "oss.2nerds": IEEE P802.1aq/D3.6 and the corresponding IETF draft (http://tools.ietf.org/html/draft-ietf-isis-ieee-aq-05) defines a series of new ISIS TLVs for the shortest-path-bridging protocol.
The attached patch file contains a simple dissector for one such TLV (the
MT-Port-Cap TLV) and several corresponding sub-TLVs (mcid, aux_mcid, digest,
and b-vid).  The digest sub-TLV dissector has not been exercised because no
suitable capture files are available (the digest sub-TLV seems not to be widely
implemented at this point in time).

Note that the codepoints mentioned in the IETF draft have changed.  The
IANA-assigned codepoints are further described in these pages:

http://www.ietf.org/mail-archive/web/spb-isis/current/msg00007.html
http://www.iana.org/assignments/isis-tlv-codepoints/isis-tlv-codepoints.xml#tlv-143

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7154

svn path=/trunk/; revision=42385