From Martin Raison via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7414 PMIPV6 support over IPV4 network
PMIPv6 is an extension of MIPv6, adding a certain number of options to the
MIPv6 protocol. These options are already implemented in the wireshark
dissector, and the MIPv6 headers are recognized, but only inside IPv6 packets.

However, RFC5844 extends the PMIPv6 protocol to make it work over IPv4 networks
: as explained in section 4 of this RFC, the MIPv6 message is carried inside an
IPv4 packet with a UDP header. The port number in the UDP header is 5436


svn path=/trunk/; revision=43505