capture_get_if_icon() is no longer platform-dependent; the platform-dependent work is now done in get_interface_type().
That means we also don't need to protect the handling of IF_DIALUP or
even IF_AIRPCAP in #ifdefs, just in case we add support for those types
in other OSes.

Nothing's uisng the capture_ethernet_16.xpm image here, so don't bother
including it.  (Does *anybody* use it?  I couldn't find any uses, but
maybe we should use that.)

svn path=/trunk/; revision=43531