Analyze->Apply as Filter and Analyze->Prepare a Filter cause crashes
user_data is NULL in the callbacks from the menu_bar, and when preparing filters we need to find the filter widget with g_object_get_data(G_OBJECT(data), E_DFILTER_TE_KEY).
The object is saved in the global variable popup_menu_object.
Use that in the filter callbacks from the menu bar.
- remove unused widget from a couple of callbacks.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7506

svn path=/trunk/; revision=43961