Make the protocol ID field an integral field with a value_string table; use that table for setting the column as well.
dissect_xdlc_control() sets the Info column; don't set it ourselves
(dissect_xdlc_control() will overwrite it if called).

That means we should always call dissect_xdlc_control(), regardless of
whether we're building a protocol tree or not.  Just get rid of the test
for that.

Append the protocol ID, if present, to the column.

svn path=/trunk/; revision=44234