Implement File→Merge.
The GTK+ and native Win32 versions are slightly different. The GTK+
version lets you select an output file type and the Win32 version uses
the existing capture filetype. We do the latter for now.

Start documenting significant UI changes in README.qt. This might be
better handled on the wiki.

svn path=/trunk/; revision=44797