*NIX (autofoo and cmake) part of the fix for https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6134 :
Make VERSION_MAJOR, VERSION_MINOR, and VERSION_MICRO available in config.h so
(e.g.) dissector writers can #if around them as necessary to make a piece of
code compile with different versions of Wireshark.  (Hopefully VERSION_MICRO
is not important in this respect!)

Windows also defines VERSION_EXTRA which can be used to help identify custom
builds; this is not done for these build methods (yet?).

svn path=/trunk/; revision=44986