Copy over: evision 44885 - Also disable the resize grip on the profile bar (as it is on the info and packets bars; this is for GTK < 3.0).
This fixes the problem I had that the profile bar normally took up way too
much (horizontal) room when I start Wireshark (causing the contents of the
much more interesting packets bar to be truncated).


Revision 45012 - Fix https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7731 :

Allow the configured number of dfilters to be added to the drop-down list
(instead of limiting the list--on a freshly installed system with no recent
dfilters--to just 2).  Regression caused by r42439.
--------------------------------------------------------------------------------
Revision 45108 - allow setting a capture comment from the statusbar when a capture was started

for the decision to allow input of a capture comment, we should make no
difference between the statusbar icon and Statistics / Summary menu item

svn path=/trunk-1.8/; revision=45171