According to
	http://home.martin.cc/linux/prism

there's a set of DID type values different from the ones we were using,
and there are captures out there that use values from both sets. 
Support both sets.

That page also says that a "status" value of 0 means "supplied"; treat
zero as meaning "supplied", and, if it's not zero for a field, don't
include it.

The "Mac Time" is, according to that page, the lower 32 bits of the MAC
timestamp; report it as such.

Fix some field names that were copied-and-pasted but not changed.

The RSSI and signal quality values are numbers, so show them in decimal.

The "signal" and "noise" values appear to be signed numbers, so make
them signed rather than unsigned and show them in decimal.

Show the data rate in the same style as it's shown in the radiotap
dissector.

Show the frame length in decimal; we probably have relatively few users
with 16 fingers.

svn path=/trunk/; revision=45545