From Michael Mann via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7768 IEEE802.11 dissector field size issue
There are a handful of fields in the IEEE802.11 dissector that are comprised of
a 16-bit value.  The hf array for these fields has the necessary masks to
correctly parse a 16-bit value, yet some of the fields were being added as 1
byte.  This patch corrects these fields with a proto_tree_add_item approach
(instead of proto_tree_add_[uint|boolean]).

svn path=/trunk/; revision=45828