Copy over: ---------------------------------------------------------------------------------------- Revision 44611 - From Teguh via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7657
Replace the last instance of gnutls_datum with gnutls_datum_t. The former
is deprecated, and the latter is already being used elsewhere in the file.
----------------------------------------------------------------------------------------
Revision 45465 - Make the SSL dissector work with GnuTLS compiled with nettle.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6869
----------------------------------------------------------------------------------------
Revision 45865 - Print the GNUTLS error message upon import failure.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7948

----------------------------------------------------------------------------------------
Revision 45866 - This patch will print the information if an
invalid string was entered. It would be better to have a button to click on in
the UAT dialog to show valid values, but I don't know how I could do that with
the UAT system. So I'm simply printing it now in the error dialog, which should
be good enough.

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7949

svn path=/trunk-1.8/; revision=45926