Handle the pseudo-trailer for the ATN extended checksums by passing in the source and destination address lengths and data, and adding them in, rather than by copying those fields to a fixed-size static buffer - that gets rid of a global variable (global variables considered harmful), and also means that we don't try to copy 24 bytes of length+address from packets that don't *have* 24 bytes of length+address (which caused exceptions to be thrown on some OSI captures I have).
Construct some LI #defines out of other LI #defines, to make it a bit
clearer why they have the values they do.

Support the "additional information related to the clearing of the
connection" variable part parameter of the COTP DR packet (which just
means giving it a name, as its contents are user-defined - some HP-UX
OSI stack appears to just stick in a string saying that it's said
stack).

Make the code that decodes the variable part of a DR packet look like
the code that decodes the variable part of most other packets.

For COTP CR packets, determine the class up front by checking whether
the length is > 2.  (At some point we might want to associate a class
indication with the COTP connection, if we see the connection setup,
and, if we have that indication, use it in preference to the
heuristics.)

Make the code to handle various length indicator values in the ATN case
more like the code in the non-ATN case.

Dissect the variable part of COTP ER packets.

Fix tpyos (TDPU->TPDU, tdpu->tpdu) and typpoes (accross->across).

Clean up white space.

svn path=/trunk/; revision=46060