Copy over: Revision 46113 - The previous change means the "len" argument is now being used. -------------------------------------------------------------------------------- Revision 46111 - Allign the tables for dissection. From me try to fix the dissection of SONtransferApplicationIdentity_PDU.
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8008

--------------------------------------------------------------------------------
Revision 46081 - - In de_bssgp_pdu_in_error() use current_offset isof 0.
- in function de_bssgp_ran_inf_error_rim_cont. Error rim
container doesn't have field rim sequence number. This field exists in the rest
of containers but not in this one. So I have removed this field. 
In this function we have field pdu in error and this field is mandatory, not
optional

https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8000

svn path=/trunk-1.8/; revision=46114