Revert SVN #46429 again (for now);
It appears that the compiler on
the OSX-10.5-x86 buildbot must be pretty old.

All the other *nix buildbot compilers do not
check the dummy param names used in the definition
of 'xxx' below to see if they are shadowing
anything.

void foo(tvbuff_t tvb, ...) {
  'void (*xxx)(tvbuff_t tvb, ....) = NULL;'
  ...
}

I do seem to remember reading that gcc once
had this issue but was fixed eventually.


svn path=/trunk/; revision=46430