Copy over: Revision 46555 - Initialize a variable to avoid a (false) may-be-used-uninitialized warning turned error. -------------------------------------------------------------------------------- Revision 46553 - Handle hop-by-hop-id collisions (when matching requests and answers) better: instead of storing the requests in an se_tree (keyed by hop-by-hop-id), store an se_tree which is itself keyed by frame number.  When looking for a matching request, first find the tree of pdus (by hop-by-hop-id) and then look for the largest frame number less than or equal to the answer's frame number.  Also verify that the end-to-end-id matches.
Move the conversation structure out of packet-diameter.h: it's only used in
packet-diameter.c

--------------------------------------------------------------------------------
Revision 46271 - Slightly less hacky way to arrange that dictionary_load() be called
before hf_base[] is initialized.
--------------------------------------------------------------------------------
Revision 46266 - Fix pretty-printing of the application ID in the protocol-details pane by
putting an "unused" variable (which was taken out in r24377) back in.  This
time add a (bogus) "use" of the variable and explain why we're doing what
we're doing.

Use VALS() (in one case anyway) and PINFO_FD_NUM.

--------------------------------------------------------------------------------
Revision 46249 - There's a macro for the Request bit.  Use it instead of hard-coded 0x80.
--------------------------------------------------------------------------------
Revision 45598 - make the variable that holds tvb_length_remaining() a signed integer
--------------------------------------------------------------------------------
Revision 45553 - Try to get rid of some of the" pedantic" warnings.
--------------------------------------------------------------------------------
Revision 44438 - Use val_to_str_const() where appropriate;
Also (for a few files):
- create/use some extended value strings;
- remove unneeded #include files;
- remove unneeded variable initialization;
- re-order fcns slightly so prefs_reg_handoff...() at end, etc

svn path=/trunk-1.8/; revision=46634