Copy over: Revision 46646 - Clean up white space.
Add a return where I presume it was intended to be - a 4-octet address
is completely handled in that if clause, so there's no reason to fall
through.

Fix a comment.
--------------------------------------------------------------------------------
Revision 46645 - Fix some valgrind warnings from the capture attached to
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8110

I can't reproduce the actual reported crash on trunk.

--------------------------------------------------------------------------------
Revision 45646 - (try to) fix https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7871

increment offset to point to the first byte after the options
(the code used to set an absolute position, if that was 0, we were stuck
 in an endless loop)
--------------------------------------------------------------------------------
Revision 45059 - replaced decode_boolean_bitfield calls with itemized filters

svn path=/trunk-1.8/; revision=46659

Conflicts:
	epan/dissectors/packet-clnp.c
	epan/osi-utils.c