From report of Elavarasan via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8099 Named field for each item in Selected Registrar Config methods is not available
The structure of both the Config Methods and Selected Registrar Config Methods are  the same.

Display filters such as "wps.config_methods.pushbutton", "wps.config_methods.display", etc are available for the different methods in the Config methods whereas the display filters for the elements in the Selected Registrar Config methods are found to be the same as the Config method Display filters.

From me :
add new filter for selected_registrar config method bitmask (and also permitted config method)

svn path=/trunk/; revision=46725