Use scrolled-windows to enclose "custom" prefs pages (User Interface, Layout, etc) in the same manner as to "protocol prefs" pages.
This ensures that the bottom row buttons won't disappear off the bottom
of the prefs window if one (or more) of the custom prefs pages is too large
to fit vertically within the available screen space.

In Wireshark 1.8, more items were added to the "User Interface" prefs page
thus leading to an increased possibility of this issue occurring.

I believe this fixes Bug #8078.

#Backport Wireshark 1.8


svn path=/trunk/; revision=46885