Rename MACOSX_VERSION to DARWIN_MAJOR_VERSION to reflect what it actually is (it's not an OS X version number, e.g. 10.7 for Leopard, it's a Darwin version number, e.g. 9 for Leopard).  Add comments before checks to indicate what we're checking for.
Put in a big long comment giving details about the issues I ran into
when trying to make this work on Leopard.

Fail if you try to run this on anything before Snow Leopard - note that
I said "trying" in the previous comment, and, yes, it was a very trying
experience....

Bump the libpng version - 1.5.12 isn't available any more.

Add some comments explaining why the code for Leopard was downloading
and installing libpng and libpixman.

Add some code to handle packages where older versions weren't compressed
with xz and newer versions were.

Add a hack to avoid finding the version of libpng that comes with X11
when trying to build Cairo on Leopard.

svn path=/trunk/; revision=47206