In Qt 4, C strings are, by default, treated as 7-bit ASCII when converted to QStrings, and, to get them to be treated as UTF-8, you need to call the setCodecForCStrings method of the QTextCodec class to set the codec to a UTF-8 codec.
In Qt 5, C strings are treated always as UTF-8 when converted to
QStrings, and there's no setCodecForCStrings method for the QTextCodec
class.

In addition, there's also no setCodecForTr method for the QTextCodec
class - and QObject isn't documented as even *having* a tr method, so I
don't know what to do there.

svn path=/trunk/; revision=47436