When decompressing, put an entry into the protocol tree that covers all the compressed data, put the entries for particular parts of that data under that entry, and, when we have a problem decompressing, add an expert info item to that entry, rather than just adding text or using proto_malformed.  Do so for *all* decompression problems; in particular, check for decompressed_entry() returning NULL, which it can do.
tvb_new_child_real_data() won't throw an exception, so don't bother
checking for that.

If, after the first pass, we can't find the uncompressed data for a
frame, report it as a dissector bug.

Check whether we have room for a byte in the uncompressed data buffer
*before* copying it.

Other cleanups.

svn path=/trunk/; revision=47922