Combine all the flags fields into one field, and get rid of the commented-out/#if-ed-out "Data1" field, as it's the same field.
Give the flags in that field names including "flags".

Shuffle the fields around to put "Flags" (a/k/a Data1) where Data1 was,
and put the bits in that field after it.

Update the URL for the spec at IBM.

Reformat the entries in hf_netb[] to use the same style.

svn path=/trunk/; revision=47933